Separate RndAnyState module
