+++ lib/wallet/src/Cardano/Wallet/Address/Book.hs
@@ -39,7 +39,7 @@ import Cardano.Wallet.Address.Derivation.Shared
 import Cardano.Wallet.Address.Derivation.SharedKey
     ( SharedKey (..) )
 import Cardano.Wallet.Primitive.Types.Address
-    ( Address (..), AddressState (..) )
+    ( Address, AddressState )
 import Data.Generics.Internal.VL
     ( Iso', iso, withIso )
 import Data.Kind
@@ -230,24 +230,6 @@ instance ( key ~ SharedKey ) => Eq (Prologue (Shared.SharedState n key)) where
 instance ( key ~ SharedKey ) => Eq (Discoveries (Shared.SharedState n key)) where
     SharedDiscoveries ext1 int1 == SharedDiscoveries ext2 int2 =
         ext1 == ext2 && int1 == int2
-
-{-------------------------------------------------------------------------------
-    HD Random address book
--------------------------------------------------------------------------------}
--- piggy-back on SeqState existing instance, to simulate the same behavior.
-instance AddressBookIso (Rnd.RndAnyState n p)
-  where
-    data Prologue (Rnd.RndAnyState n p) = PR (Prologue (Rnd.RndState n))
-    data Discoveries (Rnd.RndAnyState n p) = DR (Discoveries (Rnd.RndState n))
-
-    addressIso = withIso addressIso $ \from to ->
-        let from2 st = let (a,b) = from $ Rnd.innerState st in (PR a, DR b)
-            to2 (PR a, DR b) = Rnd.RndAnyState $ to (a,b)
-        in  iso from2 to2
-
-instance Eq (Prologue (Rnd.RndAnyState n p)) where PR a == PR b = a == b
-
-instance Eq (Discoveries (Rnd.RndAnyState n p)) where DR a == DR b = a == b
 
 -- | Isomorphism for HD random address book.
 instance AddressBookIso (Rnd.RndState n) where
