+++ lib/wallet/src/Cardano/Wallet/Address/Discovery/Random.hs
@@ -41,10 +41,8 @@ module Cardano.Wallet.Address.Discovery.Random
     , unavailablePaths
     , defaultAccountIndex
     , withRNG
+    , addDiscoveredAddress
 
-    -- ** Benchmarking
-    , RndAnyState (..)
-    , mkRndAnyState
     ) where
 import Prelude
 
@@ -80,38 +78,25 @@ import Cardano.Wallet.Primitive.Types.Address
     ( Address (..), AddressState (..) )
 import Cardano.Wallet.Primitive.Types.RewardAccount
     ( RewardAccount )
-import Cardano.Wallet.Shelley.Compatibility.Ledger
-    ( toLedger )
-import Control.Arrow
-    ( second )
 import Control.DeepSeq
     ( NFData (..) )
 import Control.Lens
     ( over )
 import Control.Monad
     ( join )
-import Data.Digest.CRC32
-    ( crc32 )
 import Data.List.NonEmpty
     ( NonEmpty (..) )
 import Data.Map
     ( Map )
-import Data.Proxy
-    ( Proxy (..) )
 import Data.Set
     ( Set )
-import Data.Word
-    ( Word32 )
 import Fmt
     ( Buildable (..), blockMapF', indentF, tupleF )
 import GHC.Generics
     ( Generic )
-import GHC.TypeLits
-    ( KnownNat, Nat, natVal )
 import System.Random
     ( RandomGen, StdGen, mkStdGen, randomR )
 
-import qualified Cardano.Wallet.Write.UTxOAssumptions as UTxOAssumptions
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map as Map
 import qualified Data.Set as Set
@@ -369,106 +354,4 @@ instance KnownAddresses (RndState n) where
             Map.foldrWithKey
                 (\path v result -> mk (toDerivationIndexes path) v : result)
                 []
-
-
-
---------------------------------------------------------------------------------
---
--- RndAnyState
---
--- For benchmarking and testing arbitrary large random wallets.
-
--- | An "unsound" alternative that can be used for benchmarking and stress
--- testing. It re-uses the same underlying structure as the `RndState` but
--- it discover addresses based on an arbitrary ratio instead of decrypting the
--- derivation path.
---
--- The proportion is stored as a type-level parameter so that we don't have to
--- alter the database schema to store it. It simply exists and depends on the
--- caller creating the wallet to define it.
-newtype RndAnyState (network :: NetworkDiscriminant) (p :: Nat) = RndAnyState
-    { innerState :: RndState network
-    } deriving (Generic, Show)
-
-instance NFData (RndAnyState n p)
-
--- | Initialize the HD random address discovery state from a root key and RNG
--- seed.
---
--- The type parameter is expected to be a ratio of addresses we ought to simply
--- recognize as ours. It is expressed in per-myriad, so "1" means 0.01%,
--- "100" means 1% and 10000 means 100%.
-mkRndAnyState
-    :: forall (p :: Nat) n. ()
-    => ByronKey 'RootK XPrv
-    -> Int
-    -> RndAnyState n p
-mkRndAnyState key seed = RndAnyState
-    { innerState = RndState
-        { hdPassphrase = payloadPassphrase key
-        , accountIndex = minBound
-        , discoveredAddresses = mempty
-        , pendingAddresses = mempty
-        , gen = mkStdGen seed
-        }
-    }
-
-instance RndStateLike (RndAnyState n p) where
-    importAddress addr (RndAnyState inner) =
-        RndAnyState <$> importAddress addr inner
-
-    addPendingAddress addr path (RndAnyState inner) =
-        RndAnyState $ addPendingAddress addr path inner
-
-    unavailablePaths (RndAnyState inner) =
-        unavailablePaths inner
-
-    defaultAccountIndex (RndAnyState inner) =
-        defaultAccountIndex inner
-
-    withRNG (RndAnyState inner) action =
-        second RndAnyState $ withRNG inner action
-
-instance KnownNat p => IsOurs (RndAnyState n p) Address where
-    isOurs addr@(Address bytes) st@(RndAnyState inner) =
-        case isOurs addr inner of
-            (Just path, inner') ->
-                (Just path, RndAnyState inner')
-
-            (Nothing, _) | crc32 bytes < p && correctAddressType ->
-                let
-                    (path, gen') = findUnusedPath
-                        (gen inner) (accountIndex inner) (unavailablePaths inner)
-
-                    inner' = addDiscoveredAddress
-                        addr Used path (inner { gen = gen' })
-                in
-                (Just (toDerivationIndexes path), RndAnyState inner')
-
-            (Nothing, _) ->
-                (Nothing, st)
-      where
-        p = floor (double (maxBound :: Word32) * double (natVal (Proxy @p)) / 10000)
-
-        double :: Integral a => a -> Double
-        double = fromIntegral
-
-        correctAddressType =
-            UTxOAssumptions.validateAddress
-                UTxOAssumptions.AllByronKeyPaymentCredentials
-                (toLedger $ Address bytes)
-
-instance IsOurs (RndAnyState n p) RewardAccount where
-    isOurs _account state = (Nothing, state)
-
-instance HasSNetworkId n => GenChange (RndAnyState n p) where
-    type ArgGenChange (RndAnyState n p) = ArgGenChange (RndState n)
-    genChange a (RndAnyState s) = RndAnyState <$> genChange a s
-
-instance CompareDiscovery (RndAnyState n p) where
-    compareDiscovery (RndAnyState s) = compareDiscovery s
-
-instance KnownAddresses (RndAnyState n p) where
-    knownAddresses (RndAnyState s) = knownAddresses s
-
