+++ lib/wallet/src/Cardano/Wallet/DB/Store/Checkpoints/Store.hs
@@ -56,6 +56,8 @@ import Cardano.Wallet.Address.Derivation.SharedKey
     ( SharedKey (..) )
 import Cardano.Wallet.Address.Discovery
     ( PendingIxs, pendingIxsFromList, pendingIxsToList )
+import Cardano.Wallet.Address.Discovery.RandomAny
+    ( Discoveries (..), Prologue (..) )
 import Cardano.Wallet.Address.Discovery.Shared
     ( CredentialType (..) )
 import Cardano.Wallet.Address.Keys.WalletKey
@@ -147,6 +149,7 @@ import UnliftIO.Exception
 
 import qualified Cardano.Wallet.Address.Derivation as W
 import qualified Cardano.Wallet.Address.Discovery.Random as Rnd
+import qualified Cardano.Wallet.Address.Discovery.RandomAny as Rnd
 import qualified Cardano.Wallet.Address.Discovery.Sequential as Seq
 import qualified Cardano.Wallet.Address.Discovery.Shared as Shared
 import qualified Cardano.Wallet.Primitive.Types as W
