+++ lib/wallet/src/Cardano/Wallet/Flavor.hs
@@ -49,7 +49,9 @@ import Cardano.Wallet.Address.Derivation.SharedKey
 import Cardano.Wallet.Address.Derivation.Shelley
     ( ShelleyKey )
 import Cardano.Wallet.Address.Discovery.Random
-    ( RndAnyState, RndState (..) )
+    ( RndState (..) )
+import Cardano.Wallet.Address.Discovery.RandomAny
+    ( RndAnyState )
 import Cardano.Wallet.Address.Discovery.Sequential
     ( SeqAnyState, SeqState )
 import Cardano.Wallet.Address.Discovery.Shared
